+++ Sources/Demo/PlaceAutocompleteDetailsViewController.swift
@@ -101,7 +101,7 @@ extension PlaceAutocompleteResultViewController {
 }
 
 extension PlaceAutocomplete.Result {
-    static let measurumentFormatter: MeasurementFormatter = {
+    static let measurementFormatter: MeasurementFormatter = {
         let formatter = MeasurementFormatter()
         formatter.unitOptions = [.naturalScale]
         formatter.numberFormatter.roundingMode = .halfUp
@@ -137,7 +137,7 @@ extension PlaceAutocomplete.Result {
             components.append(
                 (
                     name: "Estimated time",
-                    value: PlaceAutocomplete.Result.measurumentFormatter.string(from: estimatedTime)
+                    value: PlaceAutocomplete.Result.measurementFormatter.string(from: estimatedTime)
                 )
             )
         }