+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 <!-- Add changes for active work here -->
 
+- [Core] Add xcprivacy for MapboxSearch and MapboxSearchUI
+
 - [SearchUI] Update Maki icons to all SVG, latest versions from https://github.com/mapbox/maki
 - [SearchUI] Remove all custom Maki icons
 - [Unit Tests] Update and correct tests for iOS 17 using all mocked data.