+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 - [Core] Remove Swifter library dependency from MapboxSearch target (only used in Test targets)
 - [Core] Change AbstractSearchEngine.init `supportSBS: Bool = false` parameter to `apiType: ApiType = .SBS`. This changes the default API engine for discover/category and other API requests to SBS. Add ApiType enum to represent non-Autofill and non-PlaceAutocomplete SearchEngine API types.
 
+**MapboxCoreSearch**: v2.0.0-alpha.13
+
 ## 2.0.0-rc.1
 
 - [Discover] Fix charging station category canonical ID